split up tabularproject file and add in iiif browser changes
