+++ services/madoc-ts/src/extensions/projects/editors/tabular-project-custom-editor.tsx
@@ -1,17 +1,17 @@
 import React, { useMemo } from 'react';
-import type { NetConfig } from '../../../frontend/admin/components/tabular/cast-a-net/types';
-import { DynamicVaultContext } from '../../../frontend/shared/capture-models/new/DynamicVaultContext';
-import { RevisionProviderWithFeatures } from '../../../frontend/shared/capture-models/new/components/RevisionProviderWithFeatures';
-import { useCaptureModelEditorApi } from '../../../frontend/shared/capture-models/new/hooks/use-capture-model-editor-api';
-import { Revisions } from '../../../frontend/shared/capture-models/editor/stores/revisions';
-import { useLoadedCaptureModel } from '../../../frontend/shared/hooks/use-loaded-capture-model';
-import { Button, ButtonRow } from '../../../frontend/shared/navigation/Button';
-import { HrefLink } from '../../../frontend/shared/utility/href-link';
-import { useCanvasModel } from '../../../frontend/site/hooks/use-canvas-model';
-import { useCaptureModelContributionLifecycle } from '../../../frontend/site/hooks/use-capture-model-contribution-lifecycle';
-import { useProject } from '../../../frontend/site/hooks/use-project';
-import { useRouteContext } from '../../../frontend/site/hooks/use-route-context';
-import type { CanvasFull } from '../../../types/canvas-full';
+import type { NetConfig } from '@/frontend/admin/components/tabular/cast-a-net/types';
+import { DynamicVaultContext } from '@/frontend/shared/capture-models/new/DynamicVaultContext';
+import { RevisionProviderWithFeatures } from '@/frontend/shared/capture-models/new/components/RevisionProviderWithFeatures';
+import { useCaptureModelEditorApi } from '@/frontend/shared/capture-models/new/hooks/use-capture-model-editor-api';
+import { Revisions } from '@/frontend/shared/capture-models/editor/stores/revisions';
+import { useLoadedCaptureModel } from '@/frontend/shared/hooks/use-loaded-capture-model';
+import { Button, ButtonRow } from '@/frontend/shared/navigation/Button';
+import { HrefLink } from '@/frontend/shared/utility/href-link';
+import { useCanvasModel } from '@/frontend/site/hooks/use-canvas-model';
+import { useCaptureModelContributionLifecycle } from '@/frontend/site/hooks/use-capture-model-contribution-lifecycle';
+import { useProject } from '@/frontend/site/hooks/use-project';
+import { useRouteContext } from '@/frontend/site/hooks/use-route-context';
+import type { CanvasFull } from '@/types/canvas-full';
 import { TabularProjectCustomEditorCanvas } from './tabular-project-custom-editor-canvas';
 import { TabularProjectCustomEditorTable } from './tabular-project-custom-editor-table';
 import {
