+++ services/madoc-ts/src/frontend/admin/pages/crowdsourcing/projects/tabular-project/hooks/use-tabular-project-controller.ts
@@ -1,0 +1,985 @@
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import slugify from 'slugify';
+import { useMutation } from 'react-query';
+import type { Collection as IIIFCollection, InternationalString } from '@iiif/presentation-3';
+import type { TFunction } from 'i18next';
+import type { IIIFBrowserProps } from 'iiif-browser';
+import type { ApiClient } from '@/gateway/api';
+import type { CreateProject } from '@/types/schemas/create-project';
+import { buildTabularProjectSetupPayload } from '../../../../../components/tabular/cast-a-net/CastANetStructure';
+import type {
+  DefineTabularModelValue,
+  NetConfig,
+  TabularCellRef,
+  TabularModelChange,
+  TabularModelPayload,
+} from '../../../../../components/tabular/cast-a-net/types';
+import type {
+  IiifHistoryItem,
+  IiifSelectionPayload,
+  MadocCollectionSnippet,
+  MadocManifestSnippet,
+  TabularOutlineSharePayload,
+  TabularWizardStep,
+} from '../types';
+import {
+  clampToRange,
+  createIiifHomeHistoryItem,
+  createMadocHomeCollection,
+  decodeOutlinePayload,
+  encodeOutlinePayload,
+  getErrorMessage,
+  hasIntlValue,
+  internationalStringToSearchValue,
+  loadingRouteForId,
+  looksLikeManifestReference,
+  netConfigFromSharedStructure,
+  normalizeIiifInputUrl,
+  normalizeLegacyIiifKeys,
+  resolveIiifSelectionIds,
+  tabularModelValueFromSharedPayload,
+  toArray,
+  toSelectionResource,
+  toStringValue,
+  toTypeValue,
+} from '../utils';
+import { useResizableHeight } from './use-resizable-height';
+
+const STEP_DETAILS = 0;
+const STEP_SETTINGS = 1;
+const STEP_MODEL = 2;
+const STEP_NET = 3;
+const STEP_PREVIEW = 4;
+const STEP_COMPLETE = 5;
+
+const CAST_A_NET_ROWS = 5;
+const CAST_A_NET_DEFAULT_HEIGHT = 520;
+
+const SHARE_COPY_TIMEOUT = 1800;
+
+const PREVIEW_CANVAS_HEIGHT = 420;
+const PREVIEW_SPLIT_TOTAL_HEIGHT = 760;
+const PREVIEW_SPLIT_GAP = 12;
+const PREVIEW_SPLIT_DIVIDER_HEIGHT = 18;
+const PREVIEW_CANVAS_MIN_HEIGHT = 280;
+const PREVIEW_TABLE_MIN_HEIGHT = 180;
+const PREVIEW_CANVAS_MAX_HEIGHT =
+  PREVIEW_SPLIT_TOTAL_HEIGHT - PREVIEW_SPLIT_DIVIDER_HEIGHT - PREVIEW_TABLE_MIN_HEIGHT - PREVIEW_SPLIT_GAP * 2;
+
+const MAX_MADOC_COLLECTION_HOME_PAGES = 20;
+const MAX_MADOC_MANIFEST_HOME_PAGES = 40;
+const IIIF_HOME_LOCAL_STORAGE_KEY = 'iiif-browser-tabular-project-v2';
+
+export const TABULAR_PROJECT_STEP_IDS = {
+  details: STEP_DETAILS,
+  settings: STEP_SETTINGS,
+  model: STEP_MODEL,
+  net: STEP_NET,
+  preview: STEP_PREVIEW,
+  complete: STEP_COMPLETE,
+} as const;
+
+interface UseTabularProjectControllerOptions {
+  api: ApiClient;
+  defaultLocale: string;
+  siteSlug?: string;
+  t: TFunction;
+}
+
+interface ProjectMutationResult {
+  id?: string | number;
+  slug?: string;
+  error?: string;
+}
+
+export function useTabularProjectController(options: UseTabularProjectControllerOptions) {
+  const { api, defaultLocale, siteSlug, t } = options;
+  const closeBrowserRef = useRef<null | (() => void)>(null);
+
+  const [step, setStep] = useState(STEP_DETAILS);
+  const [label, setLabel] = useState<InternationalString>({ [defaultLocale]: [''] });
+  const [summary, setSummary] = useState<InternationalString>({ [defaultLocale]: [''] });
+  const [slug, setSlug] = useState('');
+  const [autoSlug, setAutoSlug] = useState(true);
+
+  const [enableZoomTracking, setEnableZoomTracking] = useState(false);
+  const [manifestId, setManifestId] = useState<string | undefined>();
+  const [canvasId, setCanvasId] = useState<string | undefined>();
+
+  const [iiifError, setIiifError] = useState<string | null>(null);
+  const [iiifBrowserSelection, setIiifBrowserSelection] = useState<string | null>(null);
+  const [iiifMadocSearchInput, setIiifMadocSearchInput] = useState('');
+  const [iiifMadocSearchQuery, setIiifMadocSearchQuery] = useState('');
+  const [iiifPasteUrl, setIiifPasteUrl] = useState('');
+  const [iiifStartupUrl, setIiifStartupUrl] = useState<string | null>(null);
+  const [iiifHomeCollection, setIiifHomeCollection] = useState<IIIFCollection | null>(null);
+  const [isLoadingIiifHome, setIsLoadingIiifHome] = useState(false);
+  const [iiifHomeLoadError, setIiifHomeLoadError] = useState<string | null>(null);
+  const [iiifBrowserModalError, setIiifBrowserModalError] = useState<string | null>(null);
+  const [iiifBrowserVersion, setIiifBrowserVersion] = useState(0);
+
+  const [tabularModel, setTabularModel] = useState<DefineTabularModelValue>({
+    columns: 0,
+    previewRows: 0,
+    headings: [],
+  });
+  const [tabularPayload, setTabularPayload] = useState<TabularModelPayload | null>(null);
+  const [isModelValid, setIsModelValid] = useState(false);
+
+  const [netConfig, setNetConfig] = useState<NetConfig>({
+    rows: CAST_A_NET_ROWS,
+    cols: 5,
+    top: 10,
+    left: 10,
+    width: 80,
+    height: 80,
+    rowPositions: [],
+    colPositions: [],
+  });
+
+  const { height: castANetHeight, startResize: startCastANetResize } = useResizableHeight(CAST_A_NET_DEFAULT_HEIGHT, {
+    min: 360,
+    max: 760,
+  });
+
+  const { height: previewCanvasHeight, startResize: startPreviewResize } = useResizableHeight(
+    clampToRange(PREVIEW_CANVAS_HEIGHT, PREVIEW_CANVAS_MIN_HEIGHT, PREVIEW_CANVAS_MAX_HEIGHT),
+    {
+      min: PREVIEW_CANVAS_MIN_HEIGHT,
+      max: PREVIEW_CANVAS_MAX_HEIGHT,
+    }
+  );
+
+  const [isCastANetDividerHover, setIsCastANetDividerHover] = useState(false);
+  const [previewActiveCell, setPreviewActiveCell] = useState<TabularCellRef | null>(null);
+  const [previewRows, setPreviewRows] = useState<string[][]>([]);
+  const [previewAdditionalRows, setPreviewAdditionalRows] = useState(0);
+  const [shareCopied, setShareCopied] = useState<'idle' | 'copied' | 'error'>('idle');
+  const [didLoadSharedOutline, setDidLoadSharedOutline] = useState(false);
+
+  const [saveProject, { status, data, isSuccess, reset }] = useMutation<ProjectMutationResult, unknown, CreateProject>(
+    async (config: CreateProject) => {
+      try {
+        return await api.createProject(config);
+      } catch (error) {
+        return { error: getErrorMessage(error, 'Unable to create project.') };
+      }
+    }
+  );
+
+  const isError = !!data?.error;
+  const isProjectCompleted = isSuccess && !isError;
+  const createdProjectPath = data?.id ? `/projects/${data.id}` : data?.slug ? `/projects/${data.slug}` : null;
+  const detailsDone = hasIntlValue(label) && hasIntlValue(summary) && !!slug.trim();
+  const hasImage = Boolean(manifestId && canvasId);
+  const requiresNetStep = enableZoomTracking && hasImage;
+  const modelSaved = tabularModel.saved ? tabularModel.saved.every(Boolean) : false;
+  const previewRowOffset = enableZoomTracking && hasImage ? 1 : 0;
+  const basePreviewTableRowCount = Math.max(1, (netConfig.rows || CAST_A_NET_ROWS) - previewRowOffset);
+  const previewTableRowCount = basePreviewTableRowCount + previewAdditionalRows;
+  const previewTableHeight = Math.max(
+    PREVIEW_TABLE_MIN_HEIGHT,
+    PREVIEW_SPLIT_TOTAL_HEIGHT - previewCanvasHeight - PREVIEW_SPLIT_DIVIDER_HEIGHT - PREVIEW_SPLIT_GAP * 2
+  );
+  const modelColumnCount = Math.max(1, tabularModel.columns || 1);
+  const netColumnCount = Math.max(1, netConfig.cols);
+  const configuredColumnCount = Math.max(1, tabularPayload?.columns?.length || modelColumnCount);
+  const primaryLabel = Object.values(label)[0]?.join('').trim() || '';
+  const primarySummary = Object.values(summary)[0]?.join('').trim() || '';
+
+  useEffect(() => {
+    if (autoSlug) {
+      const textLabel = Object.values(label)[0]?.join('') || '';
+      setSlug(slugify(textLabel, { lower: true }));
+    }
+  }, [autoSlug, label]);
+
+  useEffect(() => {
+    if (!modelSaved) {
+      return;
+    }
+
+    const colsFromSavedModel = Math.max(1, Math.floor(tabularModel.columns || tabularPayload?.columns?.length || 0));
+
+    setNetConfig(prev => ({
+      ...prev,
+      cols: colsFromSavedModel,
+      rows: CAST_A_NET_ROWS,
+    }));
+  }, [modelSaved, tabularModel.columns, tabularPayload]);
+
+  const previewColumns = useMemo(
+    () =>
+      tabularPayload?.columns?.length
+        ? tabularPayload.columns.map(column => column.label || '')
+        : Array.from({ length: modelColumnCount }, (_, index) => tabularModel.headings?.[index] ?? ''),
+    [tabularPayload, modelColumnCount, tabularModel.headings]
+  );
+
+  const previewTooltips = useMemo(
+    () =>
+      tabularPayload?.columns?.length
+        ? tabularPayload.columns.map(column => column.helpText || '')
+        : Array.from({ length: modelColumnCount }, (_, index) => tabularModel.helpText?.[index] ?? ''),
+    [tabularPayload, modelColumnCount, tabularModel.helpText]
+  );
+
+  const netHeadings = useMemo(
+    () => Array.from({ length: netColumnCount }, (_, index) => tabularModel.headings?.[index] ?? ''),
+    [netColumnCount, tabularModel.headings]
+  );
+
+  const netTooltips = useMemo(
+    () => Array.from({ length: netColumnCount }, (_, index) => tabularModel.helpText?.[index] ?? ''),
+    [netColumnCount, tabularModel.helpText]
+  );
+
+  useEffect(() => {
+    const rowCount = previewTableRowCount;
+    const colCount = Math.max(1, previewColumns.length || 1);
+
+    setPreviewRows(prev =>
+      Array.from({ length: rowCount }, (_row, rowIndex) =>
+        Array.from({ length: colCount }, (_col, colIndex) => prev[rowIndex]?.[colIndex] ?? '')
+      )
+    );
+  }, [previewTableRowCount, previewColumns.length]);
+
+  useEffect(() => {
+    if (!previewActiveCell) {
+      return;
+    }
+
+    const maxRow = previewTableRowCount - 1;
+    const maxCol = Math.max(1, previewColumns.length || 1) - 1;
+
+    if (previewActiveCell.row > maxRow || previewActiveCell.col > maxCol) {
+      setPreviewActiveCell(null);
+    }
+  }, [previewActiveCell, previewTableRowCount, previewColumns.length]);
+
+  useEffect(() => {
+    if (shareCopied !== 'copied' || typeof window === 'undefined') {
+      return;
+    }
+
+    const timeout = window.setTimeout(() => setShareCopied('idle'), SHARE_COPY_TIMEOUT);
+    return () => window.clearTimeout(timeout);
+  }, [shareCopied]);
+
+  useEffect(() => {
+    if (!siteSlug || typeof window === 'undefined') {
+      return;
+    }
+
+    let cancelled = false;
+    const query = iiifMadocSearchQuery.trim();
+    const queryLowerCase = query.toLowerCase();
+
+    const loadMadocHomeResources = async () => {
+      setIsLoadingIiifHome(true);
+      setIiifHomeLoadError(null);
+
+      try {
+        const allCollections: MadocCollectionSnippet[] = [];
+        let collectionPage = 1;
+        let collectionTotalPages = 1;
+
+        while (collectionPage <= collectionTotalPages && collectionPage <= MAX_MADOC_COLLECTION_HOME_PAGES) {
+          const response = await api.getCollections(collectionPage);
+          allCollections.push(
+            ...response.collections.map(collection => ({
+              id: collection.id,
+              label: collection.label,
+              thumbnail: collection.thumbnail,
+            }))
+          );
+          collectionTotalPages = Math.max(1, response.pagination.totalPages || 1);
+          collectionPage += 1;
+        }
+
+        const allManifests: MadocManifestSnippet[] = [];
+        let manifestPage = 1;
+        let manifestTotalPages = 1;
+
+        while (manifestPage <= manifestTotalPages && manifestPage <= MAX_MADOC_MANIFEST_HOME_PAGES) {
+          const response = await api.getManifests(manifestPage, { query: query || undefined });
+          allManifests.push(
+            ...response.manifests.map(manifest => ({
+              id: manifest.id,
+              label: manifest.label,
+              thumbnail: manifest.thumbnail,
+            }))
+          );
+          manifestTotalPages = Math.max(1, response.pagination.totalPages || 1);
+          manifestPage += 1;
+        }
+
+        if (cancelled) {
+          return;
+        }
+
+        const filteredCollections = query
+          ? allCollections.filter(collection =>
+              internationalStringToSearchValue(collection.label).includes(queryLowerCase)
+            )
+          : allCollections;
+
+        const nextHomeCollection = createMadocHomeCollection({
+          siteSlug,
+          query,
+          collections: filteredCollections,
+          manifests: allManifests,
+        });
+
+        setIiifHomeCollection(nextHomeCollection);
+        setIiifHomeLoadError(null);
+        setIiifBrowserVersion(version => version + 1);
+      } catch (error) {
+        if (cancelled) {
+          return;
+        }
+
+        setIiifHomeCollection(null);
+        setIiifHomeLoadError(getErrorMessage(error, t('Unable to load Madoc manifests and collections.')));
+      } finally {
+        if (!cancelled) {
+          setIsLoadingIiifHome(false);
+        }
+      }
+    };
+
+    void loadMadocHomeResources();
+
+    return () => {
+      cancelled = true;
+    };
+  }, [api, iiifMadocSearchQuery, siteSlug, t]);
+
+  useEffect(() => {
+    if (didLoadSharedOutline || typeof window === 'undefined') {
+      return;
+    }
+
+    setDidLoadSharedOutline(true);
+
+    const outlineValue = new URLSearchParams(window.location.search).get('outline');
+    if (!outlineValue) {
+      return;
+    }
+
+    const shared = decodeOutlinePayload(outlineValue);
+    if (!shared) {
+      setIiifError(t('The shared outline could not be loaded.'));
+      return;
+    }
+
+    if (shared.label) {
+      setLabel(shared.label);
+    }
+
+    if (shared.summary) {
+      setSummary(shared.summary);
+    }
+
+    if (typeof shared.slug === 'string') {
+      setSlug(shared.slug);
+      setAutoSlug(false);
+    }
+
+    if (typeof shared.enableZoomTracking === 'boolean') {
+      setEnableZoomTracking(shared.enableZoomTracking);
+    }
+
+    if (shared.iiif?.manifestId) {
+      setManifestId(shared.iiif.manifestId);
+    }
+
+    if (shared.iiif?.canvasId) {
+      setCanvasId(shared.iiif.canvasId);
+    }
+
+    if (shared.tabular?.model) {
+      setTabularPayload(shared.tabular.model);
+      setTabularModel(tabularModelValueFromSharedPayload(shared.tabular.model));
+      setIsModelValid(true);
+    }
+
+    if (shared.tabular?.structure) {
+      const restoredNet = netConfigFromSharedStructure(shared.tabular.structure);
+      if (restoredNet) {
+        setNetConfig(restoredNet);
+      }
+    }
+
+    if (shared.tabular?.model) {
+      setStep(STEP_PREVIEW);
+    }
+  }, [didLoadSharedOutline, t]);
+
+  const clearImageSelection = useCallback(() => {
+    setManifestId(undefined);
+    setCanvasId(undefined);
+    setIiifError(null);
+    setIiifBrowserSelection(null);
+
+    if (step === STEP_NET) {
+      setStep(STEP_SETTINGS);
+    }
+  }, [step]);
+
+  const setBrowserCloseHandler = useCallback((close?: () => void) => {
+    closeBrowserRef.current = close || null;
+  }, []);
+
+  const setupPayload = useMemo(() => {
+    if (!tabularPayload) {
+      return null;
+    }
+
+    return buildTabularProjectSetupPayload(netConfig, tabularPayload);
+  }, [netConfig, tabularPayload]);
+
+  const projectDetailsForConfirmation = useMemo(
+    () => ({
+      label,
+      summary,
+      slug,
+      enableZoomTracking,
+      iiif: {
+        manifestId: manifestId ?? null,
+        canvasId: canvasId ?? null,
+      },
+    }),
+    [label, summary, slug, enableZoomTracking, manifestId, canvasId]
+  );
+
+  const createProjectPayload = useMemo<CreateProject | null>(() => {
+    if (!setupPayload) {
+      return null;
+    }
+
+    return {
+      label,
+      summary,
+      slug,
+      template: 'tabular-project',
+      template_options: {
+        enableZoomTracking,
+        iiif: { manifestId, canvasId },
+        tabular: setupPayload,
+      },
+      template_config: {
+        enableZoomTracking,
+        iiif: { manifestId, canvasId },
+        tabular: setupPayload,
+      },
+      remote_template: null,
+    };
+  }, [label, summary, slug, enableZoomTracking, manifestId, canvasId, setupPayload]);
+
+  const shareUrl = useMemo(() => {
+    if (typeof window === 'undefined' || !setupPayload) {
+      return '';
+    }
+
+    const outline: TabularOutlineSharePayload = {
+      label,
+      summary,
+      slug,
+      enableZoomTracking,
+      iiif: { manifestId, canvasId },
+      tabular: setupPayload,
+    };
+
+    const encoded = encodeOutlinePayload(outline);
+    if (!encoded) {
+      return '';
+    }
+
+    return `${window.location.origin}${window.location.pathname}?outline=${encodeURIComponent(encoded)}`;
+  }, [label, summary, slug, enableZoomTracking, manifestId, canvasId, setupPayload]);
+
+  const canTrackPreviewOnCanvas = Boolean(
+    enableZoomTracking &&
+    hasImage &&
+    setupPayload?.structure?.columnWidthsPctOfPage?.length &&
+    setupPayload?.structure?.rowHeightsPctOfPage?.length
+  );
+
+  const previewCanvasActiveCell = useMemo<TabularCellRef | null>(() => {
+    if (!canTrackPreviewOnCanvas || !previewActiveCell) {
+      return null;
+    }
+
+    return {
+      row: previewActiveCell.row + previewRowOffset,
+      col: previewActiveCell.col,
+    };
+  }, [canTrackPreviewOnCanvas, previewActiveCell, previewRowOffset]);
+
+  const copyShareLink = useCallback(async () => {
+    if (!shareUrl) {
+      return;
+    }
+
+    if (typeof navigator === 'undefined' || !navigator.clipboard?.writeText) {
+      setShareCopied('error');
+      return;
+    }
+
+    try {
+      await navigator.clipboard.writeText(shareUrl);
+      setShareCopied('copied');
+    } catch {
+      setShareCopied('error');
+    }
+  }, [shareUrl]);
+
+  const nudgePreviewNet = useCallback((x: number, y: number) => {
+    setNetConfig(prev => {
+      const nextLeft = clampToRange(prev.left + x, 0, 100 - prev.width);
+      const nextTop = clampToRange(prev.top + y, 0, 100 - prev.height);
+      return {
+        ...prev,
+        left: nextLeft,
+        top: nextTop,
+      };
+    });
+  }, []);
+
+  const addPreviewRow = useCallback(() => {
+    setPreviewAdditionalRows(prev => prev + 1);
+  }, []);
+
+  const moveNextFromModel = useCallback(() => {
+    if (!isModelValid || !modelSaved) {
+      return;
+    }
+
+    if (requiresNetStep) {
+      setStep(STEP_NET);
+      return;
+    }
+
+    setStep(STEP_PREVIEW);
+  }, [isModelValid, modelSaved, requiresNetStep]);
+
+  useEffect(() => {
+    if (isProjectCompleted) {
+      setStep(STEP_COMPLETE);
+    }
+  }, [isProjectCompleted]);
+
+  const steps = useMemo<TabularWizardStep[]>(
+    () => [
+      { id: STEP_DETAILS, label: t('Project details') },
+      { id: STEP_SETTINGS, label: t('Additional settings') },
+      { id: STEP_MODEL, label: t('Define tabular model') },
+      { id: STEP_NET, label: t('Cast a net'), disabled: !requiresNetStep },
+      { id: STEP_PREVIEW, label: t('Preview') },
+      { id: STEP_COMPLETE, label: t('Complete') },
+    ],
+    [t, requiresNetStep]
+  );
+
+  const goToStep = useCallback(
+    (id: number, disabled?: boolean) => {
+      if (disabled) {
+        return;
+      }
+
+      if (isProjectCompleted && id !== STEP_COMPLETE) {
+        return;
+      }
+
+      if (id <= step) {
+        setStep(id);
+      }
+    },
+    [isProjectCompleted, step]
+  );
+
+  const onAddCanvas = useCallback(
+    (result: IiifSelectionPayload | IiifSelectionPayload[]) => {
+      const first = Array.isArray(result) ? result[0] : result;
+      const resource = toSelectionResource(first?.resource || first);
+      const parent = toSelectionResource(first?.parent || null);
+      const resourceId = toStringValue(resource.id);
+      const resourceType = toTypeValue(resource.type);
+      const parentId = toStringValue(parent.id);
+      const parentType = toTypeValue(parent.type);
+      const isCanvasSelection = !resourceType || resourceType === 'canvas';
+
+      if (!resourceId || !isCanvasSelection) {
+        setIiifError(t('Select a canvas from the IIIF browser.'));
+        return;
+      }
+
+      const manifestFromParent = looksLikeManifestReference(parentId, parentType) ? parentId : undefined;
+      const manifestFromPartOf = toArray(resource.partOf)
+        .map(part => toSelectionResource(part))
+        .map(part => ({ id: toStringValue(part.id), type: toTypeValue(part.type) }))
+        .find(part => looksLikeManifestReference(part.id, part.type))?.id;
+
+      const { manifestId: resolvedManifestId, canvasId: resolvedCanvasId } = resolveIiifSelectionIds({
+        manifestId: manifestFromParent || manifestFromPartOf,
+        canvasId: resourceId,
+        siteSlug,
+      });
+
+      if (!resolvedManifestId || !resolvedCanvasId) {
+        setIiifError(t('Select a canvas from within a manifest.'));
+        return;
+      }
+
+      setCanvasId(resolvedCanvasId);
+      setManifestId(resolvedManifestId);
+      setIiifError(null);
+      setIiifBrowserSelection(`${resource.id}`);
+      closeBrowserRef.current?.();
+    },
+    [siteSlug, t]
+  );
+
+  const searchMadocResources = useCallback(() => {
+    setIiifBrowserModalError(null);
+    setIiifStartupUrl(null);
+    setIiifMadocSearchQuery(iiifMadocSearchInput.trim());
+  }, [iiifMadocSearchInput]);
+
+  const clearMadocSearch = useCallback(() => {
+    setIiifBrowserModalError(null);
+    setIiifStartupUrl(null);
+    setIiifMadocSearchInput('');
+    setIiifMadocSearchQuery('');
+  }, []);
+
+  const openIiifUrl = useCallback(() => {
+    const normalizedUrl = normalizeIiifInputUrl(iiifPasteUrl);
+    if (!normalizedUrl) {
+      setIiifBrowserModalError(t('Paste a valid IIIF URL to open.'));
+      return;
+    }
+
+    setIiifBrowserModalError(null);
+    setIiifStartupUrl(normalizedUrl);
+    setIiifBrowserVersion(version => version + 1);
+  }, [iiifPasteUrl, t]);
+
+  const iiifHomeHistoryItem = useMemo(() => {
+    if (!iiifHomeCollection) {
+      return null;
+    }
+
+    return createIiifHomeHistoryItem(iiifHomeCollection);
+  }, [iiifHomeCollection]);
+
+  const navigationOptions: IIIFBrowserProps['navigation'] = useMemo(() => {
+    return {
+      clickToSelect: true,
+      doubleClickToNavigate: true,
+      clickToNavigate: false,
+      markedResources: [],
+      canSelectOnlyAllowedDomains: false,
+      multiSelect: false,
+      allowNavigationToBuiltInPages: false,
+      canNavigateToCanvas: true,
+      canNavigateToManifest: true,
+      canNavigateToCollection: true,
+      canSelectCollection: false,
+      canSelectManifest: false,
+      canSelectCanvas: true,
+    } satisfies IIIFBrowserProps['navigation'];
+  }, []);
+
+  const historyOptions: IIIFBrowserProps['history'] = useMemo(() => {
+    const initialHistory: IiifHistoryItem[] = [];
+    const normalizedStartupUrl = iiifStartupUrl ? normalizeIiifInputUrl(iiifStartupUrl) : undefined;
+
+    if (normalizedStartupUrl) {
+      initialHistory.push({
+        url: normalizedStartupUrl,
+        route: loadingRouteForId(normalizedStartupUrl),
+        resource: null,
+        timestamp: new Date().toISOString(),
+      });
+    }
+
+    if (iiifHomeHistoryItem) {
+      initialHistory.push(iiifHomeHistoryItem);
+    }
+
+    if (!initialHistory.length) {
+      initialHistory.push({
+        url: 'iiif://home',
+        route: '/',
+        resource: null,
+        timestamp: new Date().toISOString(),
+      });
+    }
+
+    return {
+      saveToLocalStorage: false,
+      restoreFromLocalStorage: false,
+      localStorageKey: IIIF_HOME_LOCAL_STORAGE_KEY,
+      initialHistory,
+      initialHistoryCursor: 0,
+      seedCollections: iiifHomeCollection ? [iiifHomeCollection] : [],
+      preprocessManifest: async manifest => {
+        const normalized = normalizeLegacyIiifKeys(manifest);
+        return normalized.value;
+      },
+    };
+  }, [iiifHomeCollection, iiifHomeHistoryItem, iiifStartupUrl]);
+
+  const rootCollection =
+    typeof window !== 'undefined' && siteSlug
+      ? `${window.location.origin}/s/${siteSlug}/madoc/api/collections/root`
+      : 'iiif://home';
+
+  const searchOptions: IIIFBrowserProps['search'] = useMemo(() => {
+    if (!siteSlug || typeof window === 'undefined') {
+      return undefined;
+    }
+
+    return {
+      enableWithinCollection: false,
+      enableExternal: true,
+      combination: {
+        mode: 'externalFirst',
+        maxExternalResults: 20,
+      },
+      typesense: {
+        host: window.location.hostname,
+        port: `/s/${siteSlug}/madoc/api/typesense`,
+        protocol: window.location.protocol.replace(':', ''),
+        path: `/s/${siteSlug}/madoc/api/typesense`,
+        collection: 'madoc_site_1',
+        searchParams: {
+          query_by: 'resource_label,search_text,metadata_label,metadata_pairs',
+          per_page: 20,
+        },
+        mapHitToResult(hit) {
+          const doc = hit.document;
+          const manifestIds = doc.manifest_ids as string[] | undefined;
+          const manifestHitId =
+            manifestIds && manifestIds.length > 0
+              ? manifestIds[0]
+              : typeof doc.manifest_id === 'string'
+                ? doc.manifest_id.split(':').pop()
+                : null;
+
+          const resourceUrl = manifestHitId
+            ? `${window.location.origin}/s/${siteSlug}/madoc/api/manifests/${manifestHitId}/export/3.0`
+            : String(doc.resource_id ?? doc.id);
+
+          const highlightSummary =
+            hit.highlights
+              ?.map(h => h.snippet?.trim())
+              .filter(Boolean)
+              .join(' · ') ?? null;
+
+          return {
+            id: String(doc.id),
+            label: String(doc.resource_label ?? doc.sort_label ?? 'Untitled'),
+            thumbnail: doc.thumbnail ? String(doc.thumbnail) : null,
+            summary: highlightSummary,
+            kind: 'external' as const,
+            resourceId: resourceUrl,
+            resourceType: 'Manifest',
+            metadata: doc,
+          };
+        },
+      },
+      externalSectionLabel: slug ? 'Madoc search results' : 'Madoc search results',
+    };
+  }, [siteSlug, slug]);
+
+  const uiOptions: IIIFBrowserProps['ui'] = useMemo(
+    () => ({
+      homeLink: rootCollection,
+      defaultPages: {
+        about: false,
+        bookmarks: false,
+        history: false,
+        viewSource: false,
+        homepage: false,
+      },
+    }),
+    [rootCollection]
+  );
+
+  const outputOptions: IIIFBrowserProps['output'] = useMemo<IIIFBrowserProps['output']>(
+    () => [
+      {
+        type: 'callback',
+        cb: onAddCanvas,
+        format: {
+          type: 'custom',
+          format(resource, parent, vault) {
+            const fromVault = (value: unknown) => {
+              try {
+                return vault.get(value as Parameters<typeof vault.get>[0]);
+              } catch {
+                return null;
+              }
+            };
+
+            return {
+              resource: fromVault(resource) || resource,
+              parent: fromVault(parent) || parent,
+            };
+          },
+        },
+        label: t('Use selected canvas'),
+        supportedTypes: ['Canvas'],
+      },
+    ],
+    [onAddCanvas, t]
+  );
+
+  const iiifHomeStats = useMemo(() => {
+    const items = (iiifHomeCollection?.items || []) as Array<{ type?: string }>;
+    const collections = items.filter(item => item?.type === 'Collection').length;
+    const manifests = items.filter(item => item?.type === 'Manifest').length;
+
+    return {
+      collections,
+      manifests,
+    };
+  }, [iiifHomeCollection]);
+
+  const onModelChange = useCallback((res: TabularModelChange) => {
+    setTabularPayload(res.payload);
+    setIsModelValid(res.isValid);
+  }, []);
+
+  const saveDetailsStep = useCallback(() => {
+    reset();
+    setStep(STEP_SETTINGS);
+  }, [reset]);
+
+  const saveSettingsStep = useCallback(() => {
+    reset();
+    setStep(STEP_MODEL);
+  }, [reset]);
+
+  const saveNetStep = useCallback(() => {
+    reset();
+    setStep(STEP_PREVIEW);
+  }, [reset]);
+
+  const savePreviewStep = useCallback(() => {
+    setStep(STEP_COMPLETE);
+  }, []);
+
+  const completeProject = useCallback(() => {
+    if (!createProjectPayload) {
+      return;
+    }
+
+    saveProject(createProjectPayload);
+  }, [createProjectPayload, saveProject]);
+
+  return {
+    stepIds: TABULAR_PROJECT_STEP_IDS,
+    step,
+    setStep,
+    steps,
+    isProjectCompleted,
+    goToStep,
+
+    label,
+    summary,
+    slug,
+    setLabel,
+    setSummary,
+    setSlug,
+    disableAutoSlug: () => setAutoSlug(false),
+    detailsDone,
+    saveDetailsStep,
+
+    enableZoomTracking,
+    setEnableZoomTracking,
+    hasImage,
+    manifestId,
+    canvasId,
+    iiifBrowserSelection,
+    iiifError,
+    clearImageSelection,
+    setBrowserCloseHandler,
+    saveSettingsStep,
+
+    tabularModel,
+    setTabularModel,
+    isModelValid,
+    modelSaved,
+    onModelChange,
+    moveNextFromModel,
+
+    requiresNetStep,
+    netConfig,
+    setNetConfig,
+    castANetHeight,
+    startCastANetResize,
+    isCastANetDividerHover,
+    setIsCastANetDividerHover,
+    netColumnCount,
+    netHeadings,
+    netTooltips,
+    saveNetStep,
+
+    shareUrl,
+    shareCopied,
+    copyShareLink,
+    canTrackPreviewOnCanvas,
+    previewCanvasHeight,
+    startPreviewResize,
+    previewCanvasActiveCell,
+    previewColumns,
+    previewTooltips,
+    previewTableRowCount,
+    previewRows,
+    setPreviewRows,
+    previewActiveCell,
+    setPreviewActiveCell,
+    previewTableHeight,
+    nudgePreviewNet,
+    addPreviewRow,
+    savePreviewStep,
+
+    iiifMadocSearchInput,
+    setIiifMadocSearchInput,
+    isLoadingIiifHome,
+    searchMadocResources,
+    clearMadocSearch,
+    iiifHomeStats,
+    iiifPasteUrl,
+    setIiifPasteUrl,
+    openIiifUrl,
+    iiifHomeLoadError,
+    iiifBrowserModalError,
+    iiifBrowserVersion,
+    navigationOptions,
+    historyOptions,
+    searchOptions,
+    outputOptions,
+    uiOptions,
+
+    createdProjectPath,
+    primaryLabel,
+    primarySummary,
+    configuredColumnCount,
+    projectDetailsForConfirmation,
+    setupPayload,
+    tabularPayload,
+    createProjectPayload,
+    isError,
+    errorMessage: data?.error,
+    status,
+    completeProject,
+
+    reset,
+  };
+}
